Move common experimental fixtures to conftest.py
